+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.5.0
 
 - Adds `aqueduct db schema` to print an application's data model.
+- Adds `aqueduct document serve` that serves the API documentation for an application.
 - Adds `--machine` flag to `aqueduct` tool to only emit machine-readable output.
 - Adds `defaultDelay` to `MockHTTPServer`. Defaults to null for no delay.
 - Adds `defaultResponse` to `MockHTTPServer`. Defaults to a 503 response instead of a 200.